Add files via upload
Sometimes the notification icons cache may get corrupted causing the icons to display incorrectly or distorted, and still show icons of uninstalled apps in the Other system tray icons (new name) or Taskbar corner overflow (old name) settings. When this happens, the notification icons needs to be deleted in the registry to reset and automatically recreate them.